+++ src/main/java/org/folio/services/ResponseHandlerUtil.java
@@ -5,7 +5,9 @@ import org.folio.rest.jaxrs.model.Errors;
 
 public final class ResponseHandlerUtil {
   private static final String HRID_ERROR_MESSAGE = "lower(f_unaccent(jsonb ->> 'hrid'::text))";
+  private static final String MATCH_KEY_ERROR_MESSAGE = "lower(f_unaccent(jsonb ->> 'matchKey'::text))";
   private static final String HRID = "HRID";
+  private static final String TABLE_NAME = "instance";
 
   private ResponseHandlerUtil() {
   }
@@ -25,13 +27,10 @@ public final class ResponseHandlerUtil {
   }
 
   private static String getErrorMessage(Object responseEntity) {
-    var errorMessage = responseEntity.toString();
     if (responseEntity instanceof Errors errors) {
-      errorMessage = errors.getErrors().get(0).getMessage();
-    } else if (responseEntity instanceof String message) {
-      errorMessage = message;
+      return errors.getErrors().get(0).getMessage();
     }
-    return errorMessage;
+    return responseEntity.toString();
   }
 
   private static Response createResponse(Response response) {
@@ -43,6 +42,13 @@ public final class ResponseHandlerUtil {
     }
   }
 
+  private static Response createResponse(Response response, String errorMessage) {
+    var transformedMessage = transformHridErrorMessage(errorMessage);
+    return Response.fromResponse(response)
+      .entity(transformedMessage)
+      .build();
+  }
+
   public static Response failedValidationResponse(Response response) {
     var entity = (Errors) response.getEntity();
     var errors = entity.getErrors();
@@ -51,5 +57,45 @@ public final class ResponseHandlerUtil {
     return Response.fromResponse(response)
       .entity(entity)
       .build();
+  }
+
+  // todo: use helper methods from PgUtil class when they become public
+  public static Response handleHridErrorInInstance(Response response) {
+    var statusCode = response.getStatus();
+
+    if (statusCode == 201) {
+      return response;
+    }
+
+    var errorMessage = getErrorMessage(response.getEntity());
+    if (errorMessage.contains(HRID_ERROR_MESSAGE)) {
+      return createResponse(response, errorMessage);
+    } else if (errorMessage.contains(MATCH_KEY_ERROR_MESSAGE)) {
+      return createMatchKeyResponse(response);
+    }
+    return response;
+  }
+
+  private static Response createMatchKeyResponse(Response response) {
+    var entity = response.getEntity().toString();
+    var matchKeyValue = extractValue(entity);
+    var remappedMessage = String.format("%s value already exists in table instance: %s",
+      MATCH_KEY_ERROR_MESSAGE, matchKeyValue);
+    return Response.fromResponse(response)
+      .entity(remappedMessage)
+      .build();
+  }
+
+  private static String transformHridErrorMessage(String errorMessage) {
+    var hridValue = extractValue(errorMessage);
+    return hridValue != null
+      ? String.format("%s value already exists in table %s: %s", HRID, TABLE_NAME, hridValue)
+      : errorMessage;
+  }
+
+  private static String extractValue(String errorMessage) {
+    var startIndex = errorMessage.indexOf("=(") + 2;
+    var endIndex = errorMessage.indexOf(")", startIndex);
+    return (startIndex > 1 && endIndex > startIndex) ? errorMessage.substring(startIndex, endIndex) : null;
   }
 }